requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
beautifulsoup4
requests
pydantic>=2.0
python-dotenv
google-generativeai
gunicorn
